Multiple files, Remove Chinese Chars, Some Test Cases
