+++ src/main.rs
@@ -1,6 +1,3 @@
-extern crate serde;
-extern crate serde_json;
-
 mod config;
 
 fn main() {
